+++ magro-cli/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+* A CLI option is added to specify clone destination.
+    + Use `clone --collection <dir>` or `clone -d <dir>`.
+
 ## [0.0.1]
 
 Initial release.